Add support for Table and HorizontalTable layouts
A Table is a modern type of Grid expressed as a list of Rows.
It avoids the need to explicitly number rows/columns
and makes the markup more closely match the actual
UI, making it easier to read and understand. Most uses
of traditional Grid would like be better expressed as a
Table.

A HorizontalTable is similar to a normal Table, except
that it's expressed as a list of Columns instead of a list
of Rows. It's expected to be used much less often than Table,
but is appropriate to use when the first column (instead of
the first row) contains labels, with each additional column
corresponding to a record.