Rename Microsoft.StandardUI.SourceGenerator -> Microsoft.StandardUI.Analyzers
Rationale: This assembly will include analyzers in the future, to
report errors in a better way. I should also be conceptually
simpler for users, with Analyzers (the traditional suffix for
analyzer projects) including both - they can always
include it and source may or may not actually get generated
depending on what the project does.